refactor: F안 Phase 2 — client.py 분할 + LLM Adapter/Port 삭제
- client.py 1182줄 → router.py + providers/{anthropic,openai,glm}.py + fallback.py + errors.py
- infrastructure/adapters/llm/ 7파일 → 재수출 스텁 (canonical: core.llm.* + core.agent.adapters.*)
- infrastructure/ports/{llm_port,agentic_llm_port}.py → 재수출 스텁 (canonical: core.llm.router)
- BillingError/UserCancelledError → core/llm/errors.py
- Agentic adapters (claude/openai/glm) → core/agent/adapters/로 이동
- CircuitBreaker + retry_with_backoff_generic → core/llm/fallback.py
- LLMClientPort + ContextVar DI → core/llm/router.py
- 전체 테스트 3181 pass, lint/mypy 0 errors
