+++ core/agent/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""Agentic LLM adapters — provider-specific implementations for AgenticLoop."""
