+++ core/domains/game_ip/nodes/synthesizer.py
@@ -18,9 +18,9 @@ from pydantic import ValidationError
 
 from core.config import get_node_model
 from core.infrastructure.ports.domain_port import get_domain_or_none
-from core.infrastructure.ports.llm_port import get_llm_json, get_llm_parsed, get_llm_tool
 from core.infrastructure.ports.tool_port import get_tool_executor
 from core.llm.prompts import SYNTHESIZER_SYSTEM, SYNTHESIZER_TOOLS_SUFFIX, SYNTHESIZER_USER
+from core.llm.router import get_llm_json, get_llm_parsed, get_llm_tool
 from core.state import (
     EvaluatorResult,
     GeodeState,
