+++ core/infrastructure/adapters/llm/agentic_registry.py
@@ -1,51 +1,11 @@
-"""Agentic adapter registry — P1 Gateway pattern.
+"""Agentic adapter registry — backward-compatible re-export.
 
-Maps provider names to adapter classes. Dynamic import for lazy loading.
-Includes cross-provider fallback map for multi-provider escalation.
+Canonical location: core.agent.adapters.agentic_registry
 """
 
-from __future__ import annotations
-
-import importlib
-import logging
-from typing import TYPE_CHECKING
-
-from core.config import ANTHROPIC_PRIMARY, OPENAI_PRIMARY
-
-if TYPE_CHECKING:
-    from core.infrastructure.ports.agentic_llm_port import AgenticLLMPort
-
-log = logging.getLogger(__name__)
-
-# Provider → "module_path:ClassName"
-_ADAPTER_MAP: dict[str, str] = {
-    "anthropic": "core.infrastructure.adapters.llm.claude_agentic_adapter:ClaudeAgenticAdapter",
-    "openai": "core.infrastructure.adapters.llm.openai_agentic_adapter:OpenAIAgenticAdapter",
-    "glm": "core.infrastructure.adapters.llm.glm_agentic_adapter:GlmAgenticAdapter",
-}
-
-# Cross-provider fallback: when a provider's chain is exhausted, try these.
-# GLM → OpenAI → Anthropic (Bug #6 fix: add Anthropic path for GLM)
-CROSS_PROVIDER_FALLBACK: dict[str, list[tuple[str, str]]] = {
-    "anthropic": [("openai", OPENAI_PRIMARY)],
-    "openai": [("anthropic", ANTHROPIC_PRIMARY)],
-    "glm": [("openai", OPENAI_PRIMARY), ("anthropic", ANTHROPIC_PRIMARY)],
-}
-
-
-def resolve_agentic_adapter(provider: str) -> AgenticLLMPort:
-    """Create an agentic adapter for the given provider.
-
-    Uses dynamic import to avoid loading unused providers.
-    """
-    entry = _ADAPTER_MAP.get(provider)
-    if entry is None:
-        # Unknown provider → default to OpenAI-compatible
-        log.warning("Unknown provider '%s', defaulting to openai adapter", provider)
-        entry = _ADAPTER_MAP["openai"]
-
-    module_path, class_name = entry.rsplit(":", 1)
-    mod = importlib.import_module(module_path)
-    cls = getattr(mod, class_name)
-    adapter: AgenticLLMPort = cls()
-    return adapter
+from core.agent.adapters.agentic_registry import (
+    CROSS_PROVIDER_FALLBACK as CROSS_PROVIDER_FALLBACK,
+)
+from core.agent.adapters.agentic_registry import (
+    resolve_agentic_adapter as resolve_agentic_adapter,
+)
