+++ core/llm/providers/__init__.py
@@ -1,0 +1,1 @@
+"""LLM provider implementations — Anthropic, OpenAI, GLM."""
