+++ core/verification/biasbuster.py
@@ -8,10 +8,9 @@ from typing import Any
 import numpy as np
 from pydantic import ValidationError
 
-from core.infrastructure.ports.llm_port import get_llm_json, get_llm_parsed, get_llm_tool
 from core.infrastructure.ports.tool_port import get_tool_executor
-from core.llm.client import maybe_traceable
 from core.llm.prompts import BIASBUSTER_SYSTEM, BIASBUSTER_USER
+from core.llm.router import get_llm_json, get_llm_parsed, get_llm_tool, maybe_traceable
 from core.state import AnalysisResult, BiasBusterResult, GeodeState
 
 log = logging.getLogger(__name__)
