+++ core/verification/guardrails.py
@@ -6,7 +6,7 @@ from typing import Any
 
 import numpy as np
 
-from core.llm.client import maybe_traceable
+from core.llm.router import maybe_traceable
 from core.state import AnalysisResult, EvaluatorResult, GeodeState, GuardrailResult
