+++ tests/test_claude_adapter.py
@@ -4,8 +4,7 @@ from __future__ import annotations
 
 from unittest.mock import patch
 
-from core.infrastructure.adapters.llm.claude_adapter import ClaudeAdapter
-from core.infrastructure.ports.llm_port import LLMClientPort
+from core.llm.router import ClaudeAdapter, LLMClientPort
 
 
 class TestClaudeAdapter:
@@ -14,7 +13,7 @@ class TestClaudeAdapter:
         adapter = ClaudeAdapter()
         assert isinstance(adapter, LLMClientPort)
 
-    @patch("core.infrastructure.adapters.llm.claude_adapter.call_llm")
+    @patch("core.llm.router.call_llm")
     def test_generate_delegates(self, mock_call):
         mock_call.return_value = "test response"
         adapter = ClaudeAdapter()
@@ -24,7 +23,7 @@ class TestClaudeAdapter:
             "system", "user", model=None, max_tokens=4096, temperature=0.3
         )
 
-    @patch("core.infrastructure.adapters.llm.claude_adapter.call_llm")
+    @patch("core.llm.router.call_llm")
     def test_generate_with_kwargs(self, mock_call):
         mock_call.return_value = "custom"
         adapter = ClaudeAdapter()
@@ -33,14 +32,14 @@ class TestClaudeAdapter:
             "s", "u", model="claude-opus", max_tokens=100, temperature=0.8
         )
 
-    @patch("core.infrastructure.adapters.llm.claude_adapter.call_llm_json")
+    @patch("core.llm.router.call_llm_json")
     def test_generate_structured_delegates(self, mock_call):
         mock_call.return_value = {"score": 4.2}
         adapter = ClaudeAdapter()
         result = adapter.generate_structured("system", "user")
         assert result == {"score": 4.2}
 
-    @patch("core.infrastructure.adapters.llm.claude_adapter.call_llm_streaming")
+    @patch("core.llm.router.call_llm_streaming")
     def test_generate_stream_delegates(self, mock_call):
         mock_call.return_value = iter(["hello", " ", "world"])
         adapter = ClaudeAdapter()
