+++ tests/test_e2e.py
@@ -12,7 +12,7 @@ from core.gateway.auth.cooldown import CooldownTracker
 from core.gateway.auth.profiles import AuthProfile, CredentialType, ProfileStore
 from core.gateway.auth.rotation import ProfileRotator
 from core.graph import compile_graph
-from core.infrastructure.ports.llm_port import LLMClientPort
+from core.llm.router import LLMClientPort
 from core.orchestration.hooks import HookEvent, HookSystem
 from core.state import GeodeState
 from core.tools.analysis import PSMCalculateTool, RunAnalystTool, RunEvaluatorTool
