+++ tests/test_llm_client.py
@@ -284,7 +284,7 @@ class TestCallLLMParsed:
         mock_client = MagicMock()
         mock_client.messages.parse.return_value = mock_response
 
-        with patch("core.llm.client.get_anthropic_client", return_value=mock_client):
+        with patch("core.llm.router.get_anthropic_client", return_value=mock_client):
             result = call_llm_parsed(
                 system="test system",
                 user="test user",
@@ -313,7 +313,7 @@ class TestCallLLMParsed:
         mock_client.messages.parse.return_value = mock_response
 
         with (
-            patch("core.llm.client.get_anthropic_client", return_value=mock_client),
+            patch("core.llm.router.get_anthropic_client", return_value=mock_client),
             pytest.raises(ValueError, match="LLM returned no structured output"),
         ):
             call_llm_parsed(
@@ -342,7 +342,7 @@ class TestCallLLMParsed:
         mock_client = MagicMock()
         mock_client.messages.parse.return_value = mock_response
 
-        with patch("core.llm.client.get_anthropic_client", return_value=mock_client):
+        with patch("core.llm.router.get_anthropic_client", return_value=mock_client):
             result = call_llm_parsed(
                 system="test",
                 user="test",
@@ -361,7 +361,7 @@ class TestMaybeTraceable:
     def test_returns_identity_when_disabled(self):
         from core.llm.client import maybe_traceable
 
-        with patch("core.llm.client.is_langsmith_enabled", return_value=False):
+        with patch("core.llm.router.is_langsmith_enabled", return_value=False):
             decorator = maybe_traceable(run_type="llm", name="test")
 
             # Should be identity — decorating a function returns same function
@@ -375,7 +375,7 @@ class TestMaybeTraceable:
 
         mock_traceable = MagicMock(return_value=lambda fn: fn)
         with (
-            patch("core.llm.client.is_langsmith_enabled", return_value=True),
+            patch("core.llm.router.is_langsmith_enabled", return_value=True),
             patch.dict("sys.modules", {"langsmith": MagicMock(traceable=mock_traceable)}),
         ):
             maybe_traceable(run_type="chain", name="test_fn")
@@ -384,7 +384,7 @@ class TestMaybeTraceable:
         from core.llm.client import maybe_traceable
 
         with (
-            patch("core.llm.client.is_langsmith_enabled", return_value=True),
+            patch("core.llm.router.is_langsmith_enabled", return_value=True),
             patch.dict("sys.modules", {"langsmith": None}),
         ):
             decorator = maybe_traceable(run_type="llm")
@@ -454,7 +454,7 @@ class TestProviderRouting:
             cache_read_input_tokens=0,
         )
 
-        with patch("core.llm.client.get_anthropic_client") as mock_get:
+        with patch("core.llm.router.get_anthropic_client") as mock_get:
             mock_client = MagicMock()
             mock_client.messages.parse.return_value = mock_response
             mock_get.return_value = mock_client
@@ -489,7 +489,7 @@ class TestProviderRouting:
         mock_glm_client = MagicMock()
         mock_glm_client.beta.chat.completions.parse.return_value = mock_response
 
-        with patch("core.llm.client._get_provider_client", return_value=mock_glm_client):
+        with patch("core.llm.router._get_provider_client", return_value=mock_glm_client):
             from core.llm.client import call_llm_parsed
 
             result = call_llm_parsed(
@@ -520,7 +520,7 @@ class TestProviderRouting:
         mock_openai_client = MagicMock()
         mock_openai_client.beta.chat.completions.parse.return_value = mock_response
 
-        with patch("core.llm.client._get_provider_client", return_value=mock_openai_client):
+        with patch("core.llm.router._get_provider_client", return_value=mock_openai_client):
             from core.llm.client import call_llm_parsed
 
             result = call_llm_parsed(
@@ -545,7 +545,7 @@ class TestProviderRouting:
         mock_glm_client = MagicMock()
         mock_glm_client.chat.completions.create.return_value = mock_response
 
-        with patch("core.llm.client._get_provider_client", return_value=mock_glm_client):
+        with patch("core.llm.router._get_provider_client", return_value=mock_glm_client):
             from core.llm.client import call_llm
 
             result = call_llm("system", "user", model="glm-5")
@@ -567,7 +567,7 @@ class TestProviderRouting:
             cache_read_input_tokens=0,
         )
 
-        with patch("core.llm.client.get_anthropic_client") as mock_get:
+        with patch("core.llm.router.get_anthropic_client") as mock_get:
             mock_client = MagicMock()
             mock_client.messages.create.return_value = mock_response
             mock_get.return_value = mock_client
